Fix inability to compare trees in crossover and mutation
Previously I was unable to come up with a test that is capable of
comparing the before and after effects of mutation. Now by building a
string of nodes and its address one can do just that, see if 1.) the
node value has changed 2.) the tree nodes addresses have changed. If
either is true then the tree has indeed changed and therefore is no
longer equals.