Add common region point crossover
This common region point crossover is different than the crossover Koza
outlined. instead of choosing a random point in both parent 1 and 2,
the common region aligns the two parents up, and traverses down the
tree to check if two trees have the same shape. iI terms of function
nodes, the arity has to be the same. Those nodes that have the same
arity are added to common regions to be considered for crossover.