Fix selection and reproduction assumptions
Had the concept of selection and reproduction wrong. Before I thought I
the number of parents I had to select were half of the max, however
that is not the case, what I had to do was select parents to fill up
the new population, and perform reproduction on the whole full
population.