Add tree_evaluation_3
Inspired by how I evaluate Cartesian Graphs with CGP, I have created a
third tree evaluator. It works similar to the first tree evaluator, in
that it traverses that tree.program stack in posfix order, however
instead of traversing the tree n-times where n is the number of data
rows, it only traverses the tree ONCE. The results from different stages
of the tree are recorded as an array.

Preliminary results show that this third tree evaluator is comparable to
the closure technique, evaluating 10000 trees takes ~6 seconds as
suppose to ~4 with closure technique. (Tests were done on a MacbookPro
Retina 2012)